creacion de tablas y seeder

+++ database/migrations/2020_07_25_155204_create_descuento_table.php
@@ -15,7 +15,7 @@ class CreateDescuentoTable extends Migration
     {
         Schema::create('descuento', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('descuento');
+            $table->decimal('descuento');
             $table->timestamps();
         });
     }
